techpack: audio: aw87319: fix duplicate declaration
gto uses two different kernel drivers for the left and right speaker.

This would have been fine as a module, but now that it is being built
inline, the ODM's lazy implementation causes duplicate declarations (as
they're basically the same file but for left and right.)

Just make the offending declarations static to fix the problem.

Change-Id: I258405611cb3e9235155ecc20ab36172a4d1f55a
Signed-off-by: Han Sol Jin <hansol@hansol.ca>